+++ src/test/java/com/github/jlangch/venice/impl/functions/StringFunctionsTest.java
@@ -579,6 +579,28 @@ public class StringFunctionsTest {
 	}
 
 	@Test
+	public void test_str_trim_left() {
+		final Venice venice = new Venice();
+		
+		assertEquals("abcdef", venice.eval("(str/trim-left \"abcdef\")"));
+		assertEquals("abcdef  ", venice.eval("(str/trim-left \"  abcdef  \")"));
+		assertEquals("abcdef", venice.eval("(str/trim-left \"  abcdef\")"));
+		assertEquals("abcdef  ", venice.eval("(str/trim-left \"abcdef  \")"));
+		assertEquals("", venice.eval("(str/trim-left \"  \")"));
+	}
+
+	@Test
+	public void test_str_trim_right() {
+		final Venice venice = new Venice();
+		
+		assertEquals("abcdef", venice.eval("(str/trim-right \"abcdef\")"));
+		assertEquals("  abcdef", venice.eval("(str/trim-right \"  abcdef  \")"));
+		assertEquals("  abcdef", venice.eval("(str/trim-right \"  abcdef\")"));
+		assertEquals("abcdef", venice.eval("(str/trim-right \"abcdef  \")"));
+		assertEquals("", venice.eval("(str/trim-right \"  \")"));
+	}
+
+	@Test
 	public void test_str_truncate() {
 		final Venice venice = new Venice();