+++ ChangeLog.md
@@ -16,6 +16,7 @@ and this project adheres to
 - namespace aliases (`ns-alias`, `ns-aliases`, `ns-unalias`)
 - `Object` protocol to support customized 'compareTo' for custom datatypes
 - a parsatron expression parser example 'doc/example/scripts/expr-parser.venice'
+- functions `str/trim-left` and `str/trim-right`
 
 ### Changed